Move template querying code to module
